Commit work908
Compare the control performances of agents with policy001 and with policy002.
With policy001, especially when alphaTemp has fairly large number
or when nStepGradient is large, standard deviations of agent output go to the extremely large number.

Since policy002 has the function of clipping,
it's expected that this issues can be mitigated.